Fix retrieval edge cases and PT truth radius unit bug

+++ __utils.py
@@ -160,71 +160,21 @@ def default_parameters():
     return param
 
 
-def read_parfile(param, parfile=None, json_format=False):
+def read_parfile(param, parfile=None):
     cwd = os.getcwd()
     if parfile is None:
-            print('No parameter file provided. A standard parameter file will be used.')
-            pass
-    else:
-        if json_format:
-            with open(parfile, 'r') as f:
-                paramdata = json.load(f)
-            for key, value in paramdata.items():
-                param[key] = value
-            del paramdata
-        else:
-            #print('Reading parfile: "' + parfile + '"')
-            with open(cwd + '/' + parfile, 'r') as file:
-                paramfile = file.readlines()
-            for i in paramfile:
-                if i[0] == '%' or i[0] == '\n':
-                    pass
-                else:
-                    paramline = list(i.split('\t'))
-                    paramline[-1] = paramline[-1][:-1]
-                    if len(paramline) >= 2:
-                        try:
-                            param[paramline[0]] = float(paramline[-1])
-                        except ValueError:
-                            if str(paramline[1]) == str(True):
-                                param[paramline[0]] = bool(paramline[1])
-                            elif str(paramline[1]) == str(False):
-                                param[paramline[0]] = bool("")
-                            elif str(paramline[1]) == str(None):
-                                param[paramline[0]] = None
-                            else:
-                                param[paramline[0]] = str(paramline[1])
+        raise ValueError('A parameter file path must be provided.')
 
-                        if paramline[0] == 'file_output_name':
-                            try:
-                                param[paramline[0]] = str(int(paramline[1]))
-                            except ValueError:
-                                param[paramline[0]] = str(paramline[1])
-                    else:
-                        paramline = str(paramline[0]).split()
-                        if paramline[0] == 'mol':
-                            param[paramline[0]] = paramline[1].split(',')
-                        elif paramline[0] == 'mol_vmr' or paramline[0] == 'range_mol':
-                            param[paramline[0]] = paramline[1].split(',')
-                            for ob in range(0, len(param[paramline[0]])):
-                                param[paramline[0]][ob] = float(param[paramline[0]][ob])
-                            if paramline[0] == 'mol_vmr':
-                                for num, mol in enumerate(param['mol']):
-                                    param['vmr_' + mol] = param['mol_vmr'][num]
-                            else:
-                                pass
-                        else:
-                            try:
-                                param[paramline[0]] = float(paramline[-1])
-                            except ValueError:
-                                if str(paramline[1]) == str(True):
-                                    param[paramline[0]] = bool(paramline[1])
-                                elif str(paramline[1]) == str(False):
-                                    param[paramline[0]] = bool("")
-                                elif str(paramline[1]) == str(None):
-                                    param[paramline[0]] = None
-                                else:
-                                    param[paramline[0]] = str(paramline[1])
+    if os.path.isabs(parfile):
+        parfile_path = parfile
+    else:
+        parfile_path = os.path.join(cwd, parfile)
+
+    with open(parfile_path, 'r') as f:
+        paramdata = json.load(f)
+    for key, value in paramdata.items():
+        param[key] = value
+    del paramdata
 
     param['wkg_dir'] = cwd + '/'
     if param['output_directory'] is not None:
@@ -235,9 +185,9 @@ def read_parfile(param, parfile=None, json_format=False):
     else:
         param['out_dir'] = param['wkg_dir']
 
-    src = os.path.join(cwd, parfile)
-    dst = os.path.join(param['out_dir'], os.path.basename(parfile))
-    if not os.path.exists(dst):
+    src = os.path.abspath(parfile_path)
+    dst = os.path.abspath(os.path.join(param['out_dir'], os.path.basename(parfile_path)))
+    if src != dst and not os.path.exists(dst):
         shutil.copy2(src, dst)
 
     return param
@@ -416,8 +366,10 @@ def load_input_spectrum(param):
         try:
             if param['obs_numb'] is None:
                 spectrum = np.loadtxt(param['wkg_dir'] + param['spectrum'])
+                if spectrum.ndim == 1:
+                    spectrum = spectrum.reshape(1, -1)
                 param['spectrum'] = {}
-                if len(spectrum[0, :]) == 3:
+                if spectrum.shape[1] == 3:
                     param['spectrum']['wl'] = spectrum[:, 0]            # wavelength in micron
                     param['spectrum']['Fplanet'] = spectrum[:, 1]       # (W/m2) or contrast ratio
                     param['spectrum']['error_p'] = spectrum[:, 2]       # (W/m2) or contrast ratio
@@ -438,8 +390,10 @@ def load_input_spectrum(param):
                 min_wl, max_wl = [], []
                 for obs in range(0, int(param['obs_numb'])):
                     spectrum = np.loadtxt(param['wkg_dir'] + param['spectrum' + str(obs)])
+                    if spectrum.ndim == 1:
+                        spectrum = spectrum.reshape(1, -1)
                     param['spectrum'][str(obs)] = {}
-                    if len(spectrum[0, :]) == 3:
+                    if spectrum.shape[1] == 3:
                         param['spectrum'][str(obs)]['wl'] = spectrum[:, 0]
                         param['spectrum'][str(obs)]['Fplanet'] = spectrum[:, 1]
                         param['spectrum'][str(obs)]['error_p'] = spectrum[:, 2]
